Создает контроллер. Выносит в него логику рендера карточек, кнопки showMore, попапов
