+++ src/components/abstractComponent.js
@@ -1,4 +1,4 @@
-import {createElement} from '../components/utils/render.js';
+import {createElement} from '../utils/render.js';
 
 export default class AbstractComponent {
   constructor() {
