+++ src/mock/userRank.js
@@ -1,4 +1,4 @@
-import {getRandomIntegerNumber} from '../components/utils/common.js';
+import {getRandomIntegerNumber} from '../utils/common.js';
 
 const MAX_USER_RANK = 100;
